Push a new branch first in the pull-push loop
In order to handle multiple jobs running in parallel, and thus
pushing changes in parallel, a loop running pull and then push
tries to get a job's changes pushed to the data repository.

The problem here is that if the data branch was created locally
as a new branch, then the pull before the push will fail and the
whole script will exit. See issue #33.

This is a very simple attempt to solve this issue by mimicing
a developer's workflow, who tried to push and got rejected because
of the remote having changes from another push. The lazy developer
woud then pull and try to push again.

This does look like overkill, to be honest, and is only a quick
and dirty attempt at a solution. It is not very robust because
it relies on the error output of `git push` which might change.
Unfortunately, git commands do not have distinctive error codes
that could be used instead of guessing from the error message.