Fix exit on errexit when return code is used for checks
After a rework of the logic that checks if the data branch exists
at the beginning of the script, the return code of `git ls-remote`
is used in a check while `errexit` is active. Which means that the
script will exit and the line checking the return code is never
used.
I might be missing something here, though, because I don't understand
how this ever worked, but PR #32 was declared to be tested and working.